[Optimize] Support constructing Element Tree according to the ElementTemplate info in the PreParse stage.
Add the `GreedyConstructElements` method and `TryGetElements` method to `LynxTemplateBundle`, designed respectively to create Elements and retrieve the creation results of Elements.

Add the `GreedyConstructElements` method to `LynxBinaryReader`. Execute `GreedyDecodeElementTemplateSection` followed by an immediate call to `GreedyConstructElements` to construct the tree in advance.

Add the `ConstructElement` and `TryGetElements` methods to `ParallelParseTaskScheduler`. The `ConstructElement` method supports synchronous or asynchronous construction of Elements. The `TryGetElements` method will attempt to retrieve Elements and is designed to be a thread-safe method.

In the `TemplateEntry::ElementFromBinary` method, attempt to call `template_bundle_.TryGetElements(key);` to retrieve the results of the pre-created Elements.

issue:m-6083199326
AutoSubmit:True